+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/svs/annotation/SVDuplicationAnnotationBuilder.java
@@ -36,18 +36,11 @@ public class SVDuplicationAnnotationBuilder extends StructuralVariantAnnotationB
 
 		GenomeInterval affected = variant.getAffectedIntervalOuter().withStrand(Strand.FWD);
 
-		// Case of transcript or exon loss/truncation
-		if (variant.getAffectedIntervalOuter().contains(transcript.getTXRegion()))
-			effects.add(VariantEffect.TRANSCRIPT_ABLATION);
-		if (effects.isEmpty()) {
-			for (GenomeInterval exon : transcript.getExonRegions())
-				if (affected.contains(exon))
-					effects.add(VariantEffect.EXON_LOSS_VARIANT);
-				else if (affected.overlapsWith(exon))
-					effects.add(VariantEffect.FEATURE_TRUNCATION);
-		}
-		if (!effects.isEmpty())
+		// Consider transcript amplification
+		if (variant.getAffectedIntervalOuter().overlapsWith(transcript.getTXRegion())) {
 			effects.add(VariantEffect.TRANSCRIPT_AMPLIFICATION);
+			effects.add(VariantEffect.COPY_NUMBER_GAIN);
+		}
 
 		// Case of upstream/downstream/intergenic
 		if (effects.isEmpty() && so.overlapsWithDownstreamRegion(affected))
@@ -59,6 +52,7 @@ public class SVDuplicationAnnotationBuilder extends StructuralVariantAnnotationB
 
 		// Variant is SV in any case and affects coding/non-coding tx
 		effects.add(VariantEffect.STRUCTURAL_VARIANT);
+		effects.add(VariantEffect.COPY_NUMBER_INCREASE);
 		if (transcript.isCoding())
 			effects.add(VariantEffect.CODING_TRANSCRIPT_VARIANT);
 		else