+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## jannovar-core
 
 * less intrusive escaping in `ANN` field
-* adding `coyp_number_change`, `copy_number_increase`, `copy_number_decrease`, and `transcript_amplification`, `feature_amplification` to `VariantEffect`.
+* adding `coyp_number_change`, `copy_number_increase`, `copy_number_decrease`, and `transcript_amplification`, `feature_amplification`, and others to `VariantEffect`.
 
 ## jannovar-htsjdk